fix(xpath): (long long)nd in month-shift date prints (MSVC va_arg)
Third occurrence of the width/specifier class: the day-of-month
is an int but printed with %02lld — MSVC's va_arg pulled 8 bytes
and printed garbage days (1970-01-2113123909633Z). Cast added.

+++ src/leptris/xpath/functions_ext31.c
@@ -2444,11 +2444,11 @@ int leptris_dt_shift_months(const char* ds, double delta_months,
         }
         snprintf(buf, cap, "%s%04lld-%02lld-%02lldT%02d:%02d:%s%s",
                  ny < 0 ? "-" : "", ny < 0 ? -ny : ny,
-                 nm + 1, nd, h, mi, secs, zone);
+                 nm + 1, (long long)nd, h, mi, secs, zone);
     } else {
         snprintf(buf, cap, "%s%04lld-%02lld-%02lld%s",
                  ny < 0 ? "-" : "", ny < 0 ? -ny : ny,
-                 nm + 1, nd, zone);
+                 nm + 1, (long long)nd, zone);
     }
     return 1;
 }
